Refactor: Component5
