+++ index.html
@@ -9,8 +9,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <p>test change commit</p>
-    <p>test 2 modify change commit</p>
     <script type="module" src="/src/main.jsx"></script>
   </body>
 </html>
