+++ src/App.jsx
@@ -1,16 +1,17 @@
-import React, { useState, useContext } from "react";
+import React, { useState, createContext } from "react";
 import Component1 from "./components/Component1";
-import { NameContext } from "./NameContext";
 
+// bez sensu takiego jednolinijkowca exportować do osobnego pliku
+export const NameContext = createContext("Starting name");
 
 export default function App() {
-  const [name, setName] = useState('Andrzej')
+  const [name, setName] = useState("Andrzej");
 
   return (
-    <NameContext.Provider value={{name, setName}}>
+    <NameContext.Provider value={{ name, setName }}>
       <main>
-          <h1>hello {name}</h1>
-          <Component1 />
+        <h1>hello {name}</h1>
+        <Component1 />
       </main>
     </NameContext.Provider>
   );
