+++ src/components/Component1.jsx
@@ -1,12 +1,12 @@
 import React from 'react'
-import Component3 from './Component3'
+import Component2 from './Component2'
 
 
 export default function Component1(){
   return (
   <div>
     <h2>Component1</h2>
-    <Component3 />
+    <Component2 />
   </div>
   )
 }
