+++ src/components/Component2.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import Component3 from './Component3'
+
+
+export default function Component2(){
+  return (
+  <div>
+    <h2>Component2</h2>
+    <Component3 />
+  </div>
+  )
+}
