+++ src/components/Component4.jsx
@@ -5,7 +5,7 @@ import Component5 from './Component5'
 export default function Component4(){
   return (
   <div>
-    <h2>Component4 Modified from master</h2>
+    <h2>Component4</h2>
     <Component5 />
   </div>
   )
